# api/gateway.openapi.json
{
  "components": {
    "parameters": {
      "Backend": {
        "description": "Gateway selector for one exact backend id.",
        "in": "query",
        "name": "backend",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "Cluster": {
        "description": "Gateway selector. Use a concrete cluster for writes, admission, queries, and SSE; omit or use all for fan-out GET endpoints.",
        "in": "query",
        "name": "cluster",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "Limit": {
        "description": "Maximum number of items to return.",
        "in": "query",
        "name": "limit",
        "required": false,
        "schema": {
          "default": 100,
          "minimum": 1,
          "type": "integer"
        }
      },
...
    },
    "responses": {
      "Error": {
        "content": {
          "application/json": {
            "schema": {
              "properties": {
                "error": {
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              },
              "type": "object"
            }
          }
        },
        "description": "Error response"
      }
    }
  },
  "info": {
    "description": "Admission control, prefix-cache residency, KV-event observability, and configuration API.",
    "title": "ucloud-kv-indexer Gateway API",
    "version": "0.2.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/clusters": {
      "get": {
        "description": "Lists configured clusters.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Cluster"
          },
          {
            "$ref": "#/components/parameters/Backend"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "List clusters",
        "tags": [
          "Config"
        ]
      },
      "post": {
        "description": "Creates or replaces a cluster config.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Cluster"
          },
          {
            "$ref": "#/components/parameters/Backend"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "cluster_id": {
                    "description": "Cluster id.",
                    "type": "string"
                  },
                  "display_name": {
                    "description": "Human-readable name.",
                    "type": "string"
                  },
                  "enabled": {
                    "description": "Whether the cluster is enabled.",
                    "type": "boolean"
                  },
                  "environment": {
                    "description": "Environment label.",
                    "type": "string"
                  },
                  "labels": {
                    "additionalProperties": true,
                    "description": "Free-form labels.",
                    "type": "object"
                  },
                  "maintenance_mode": {
                    "description": "Whether the cluster is in maintenance mode.",
                    "type": "boolean"
                  },
                  "region": {
                    "description": "Region label.",
                    "type": "string"
                  }
                },
                "required": [
                  "cluster_id"
                ],
                "type": "object"
              }
            }
          },
          "description": "Cluster",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "Create or update cluster",
        "tags": [
          "Config"
        ]
      }
    },
...
    "/clusters/{id}": {
      "patch": {
        "description": "Hot-updates mutable cluster fields.",
        "parameters": [
          {
            "description": "Resource identifier.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/Cluster"
          },
          {
            "$ref": "#/components/parameters/Backend"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "display_name": {
                    "description": "Human-readable name.",
                    "type": "string"
                  },
                  "enabled": {
                    "description": "Whether the cluster is enabled.",
                    "type": "boolean"
                  },
                  "labels": {
                    "additionalProperties": true,
                    "description": "Free-form labels.",
                    "type": "object"
                  },
                  "maintenance_mode": {
                    "description": "Whether the cluster is in maintenance mode.",
                    "type": "boolean"
                  }
                },
                "type": "object"
              }
            }
          },
          "description": "Cluster patch",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "Patch cluster",
        "tags": [
          "Config"
        ]
      }
    },
    "/config/audit-log": {
      "get": {
        "description": "Lists configuration mutations and profile version bumps.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Cluster"
          },
          {
            "$ref": "#/components/parameters/Backend"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "Config audit log",
        "tags": [
          "Observability"
        ]
      }
    },
    "/config/effective-policy/preview": {
      "post": {
        "description": "Evaluates the priority-ordered rule list for a request shape. Rule conditions are AND clauses; rules are OR by priority.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Cluster"
          },
          {
            "$ref": "#/components/parameters/Backend"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "cluster_id": {
                    "description": "Cluster scope to preview.",
                    "type": "string"
                  },
                  "fresh": {
                    "description": "Whether KV event signals are trusted for this preview. Defaults true.",
                    "type": "boolean"
                  },
                  "has_candidates": {
                    "description": "Whether at least one engine can serve the request. Defaults true.",
                    "type": "boolean"
                  },
                  "hash_supported": {
                    "description": "Whether the request features are hash-supported. Defaults true.",
                    "type": "boolean"
                  },
                  "hit_ratio": {
                    "description": "Synthetic KV hit ratio for previewing require_cache_hit actions.",
                    "type": "number"
                  },
                  "input_tokens": {
                    "description": "Input token count to test against token-count conditions.",
                    "type": "integer"
                  },
                  "model_id": {
                    "description": "Model scope to preview.",
                    "type": "string"
                  },
                  "tenant_id": {
                    "description": "Tenant/customer/workspace scope to preview.",
                    "type": "string"
                  },
                  "tokenized": {
                    "description": "Whether tokenization succeeded for this preview. Defaults true.",
                    "type": "boolean"
                  }
                },
                "type": "object"
              }
            }
          },
          "description": "Rule preview request",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "Preview admission rules",
        "tags": [
          "Policies"
        ]
      }
    },
    "/config/versions": {
      "get": {
        "description": "Returns the current config version. On gateway, returns one item per backend.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Cluster"
          },
          {
            "$ref": "#/components/parameters/Backend"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "Config versions",
        "tags": [
          "Observability"
        ]
      }
    },
    "/decisions": {
      "get": {
        "description": "Returns recent admission verdicts with reason, hit ratio, target, and config version.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Cluster"
          },
          {
            "$ref": "#/components/parameters/Backend"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "Recent admission decisions",
        "tags": [
          "Observability"
        ]
      }
    },
    "/engines": {
      "get": {
        "description": "Lists registered inference engines.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Cluster"
          },
          {
            "$ref": "#/components/parameters/Backend"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "List engines",
        "tags": [
          "Config"
        ]
      }
    },
    "/engines/register": {
      "post": {
        "description": "Registers or replaces an inference engine and starts its KV-event listener if configured.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Cluster"
          },
          {
            "$ref": "#/components/parameters/Backend"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "api_endpoint": {
                    "description": "OpenAI-compatible serving endpoint.",
                    "type": "string"
                  },
                  "cluster_id": {
                    "description": "Owning cluster id.",
                    "type": "string"
                  },
                  "dp_ranks": {
                    "description": "Number of data-parallel ranks.",
                    "type": "integer"
                  },
                  "draining": {
                    "description": "Whether this engine should stop receiving new work.",
                    "type": "boolean"
                  },
                  "enabled": {
                    "description": "Whether this engine is eligible.",
                    "type": "boolean"
                  },
                  "engine_id": {
                    "description": "Engine id.",
                    "type": "string"
                  },
                  "framework": {
                    "description": "Serving framework, for example vllm or sglang.",
                    "type": "string"
                  },
                  "healthy": {
                    "description": "Manual health flag.",
                    "type": "boolean"
                  },
                  "kv_event_endpoint": {
                    "description": "ZMQ KV-event endpoint for prefix-cache awareness.",
                    "type": "string"
                  },
                  "labels": {
                    "additionalProperties": true,
                    "description": "Free-form labels.",
                    "type": "object"
                  },
                  "max_model_len": {
                    "description": "Model context length.",
                    "type": "integer"
                  },
                  "max_num_seqs": {
                    "description": "Max active sequences.",
                    "type": "integer"
                  },
                  "replay_endpoint": {
                    "description": "Optional KV replay endpoint.",
                    "type": "string"
                  },
                  "served_models": {
                    "description": "Model ids served by this engine.",
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "tokenizer_endpoint": {
                    "description": "Tokenizer endpoint. Defaults to api_endpoint when omitted.",
                    "type": "string"
                  },
                  "topic": {
                    "description": "ZMQ topic, default kv-events.",
                    "type": "string"
                  }
                },
                "required": [
                  "engine_id",
                  "cluster_id",
                  "framework",
                  "served_models"
                ],
                "type": "object"
              }
            }
          },
          "description": "Engine",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "Register engine",
        "tags": [
          "Config"
        ]
      }
    },
    "/engines/unregister": {
      "post": {
        "description": "Removes an engine by engine_id.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Cluster"
          },
          {
            "$ref": "#/components/parameters/Backend"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "engine_id": {
                    "description": "Engine identifier.",
                    "type": "string"
                  }
                },
                "required": [
                  "engine_id"
                ],
                "type": "object"
              }
            }
          },
          "description": "Engine unregister request",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "Unregister engine",
        "tags": [
          "Config"
        ]
      }
    },
    "/engines/{id}": {
      "patch": {
        "description": "Hot-updates mutable engine state such as enabled, draining, health, and queue depth.",
        "parameters": [
          {
            "description": "Resource identifier.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/Cluster"
          },
          {
            "$ref": "#/components/parameters/Backend"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "draining": {
                    "description": "Whether this engine should stop receiving new work.",
                    "type": "boolean"
                  },
                  "enabled": {
                    "description": "Whether this engine is eligible.",
                    "type": "boolean"
                  },
                  "healthy": {
                    "description": "Manual health flag.",
                    "type": "boolean"
                  },
                  "labels": {
                    "additionalProperties": true,
                    "description": "Free-form labels.",
                    "type": "object"
                  },
                  "max_num_seqs": {
                    "description": "Max active sequences.",
                    "type": "integer"
                  },
                  "queue_depth": {
                    "description": "Current serving queue depth.",
                    "type": "integer"
                  }
                },
                "type": "object"
              }
            }
          },
          "description": "Engine patch",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "Patch engine",
        "tags": [
          "Config"
        ]
      }
    },
    "/event-streams": {
      "get": {
        "description": "Per-engine ZMQ listener connection, sequence, gap, decode, and queue health.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Cluster"
          },
          {
            "$ref": "#/components/parameters/Backend"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "List KV-event listener health",
        "tags": [
          "Observability"
        ]
      }
    },
...
    "/index/stats": {
      "get": {
        "description": "Lists request_key, bridge, engine, and last-event counts per namespace.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Cluster"
          },
          {
            "$ref": "#/components/parameters/Backend"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "Residency index stats",
        "tags": [
          "Observability"
        ]
      }
    },
    "/kv-events/recent": {
      "get": {
        "description": "Returns recent decoded ZMQ KV-cache events. Query parameter limit defaults to 100.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Cluster"
          },
          {
            "$ref": "#/components/parameters/Backend"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "Recent decoded KV events",
        "tags": [
          "Observability"
        ]
      }
    },
    "/kv-events/stream": {
      "get": {
        "description": "Server-Sent Events stream of decoded KV-cache events. Select exactly one backend or cluster through the gateway.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Cluster"
          },
          {
            "$ref": "#/components/parameters/Backend"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "Live decoded KV-event stream",
        "tags": [
          "Observability"
        ]
      }
    },
...
    "/openapi.json": {
      "get": {
        "description": "Returns this OpenAPI 3 document.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "OpenAPI document",
        "tags": [
          "System"
        ]
      }
    },
...
    "/query-prefix": {
      "post": {
        "description": "Mooncake/Dynamo-style per-instance prefix hit query. Accepts model plus token_ids, or model plus prompt.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Cluster"
          },
          {
            "$ref": "#/components/parameters/Backend"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "block_size": {
                    "description": "Optional override for profile block size.",
                    "type": "integer"
                  },
                  "model": {
                    "description": "Model name used to resolve profile/namespace.",
                    "type": "string"
                  },
                  "prompt": {
                    "description": "Plain prompt to tokenize when token_ids are not provided.",
                    "type": "string"
                  },
                  "tenant_id": {
                    "description": "Optional tenant id.",
                    "type": "string"
                  },
                  "token_ids": {
                    "description": "Token ids to query. If omitted, prompt is tokenized by the engine.",
                    "items": {
                      "format": "int32",
                      "type": "integer"
                    },
                    "type": "array"
                  }
                },
                "required": [
                  "model"
                ],
                "type": "object"
              }
            }
          },
          "description": "Prefix query request",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "Query prefix cache hits",
        "tags": [
          "Prefix"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
}